+++ src/Greenleaf/Router/Caching.php
@@ -14,4 +14,5 @@ use DecodeLabs\Greenleaf\Router;
 interface Caching extends Router
 {
     public function clearCache(): void;
+    public function rebuildCache(): void;
 }